requirements.txt
# Core ADK and Google Cloud
google-adk==1.18.0
google-cloud-firestore==2.21.0
google-cloud-documentai==3.7.0
google-cloud-vision==3.11.0
google-cloud-storage==3.6.0
google-generativeai==0.8.5

# API Framework
fastapi==0.121.2
uvicorn[standard]==0.38.0
python-multipart==0.0.20

# Data Validation
pydantic==2.12.4
pydantic-settings==2.12.0

# Authentication & Environment
google-auth==2.43.0
python-dotenv==1.2.1

# PDF Processing
PyPDF2==3.0.1
pdf2image==1.17.0
pillow==12.0.0

# Utilities
aiofiles==25.1.0
httpx==0.28.1
python-jose[cryptography]==3.5.0

# Testing
pytest==9.0.1
pytest-asyncio==1.3.0
pytest-cov==4.1.0

# Monitoring & Logging
python-json-logger==4.0.0

# Data Handling
pandas>=2.0.0
openpyxl>=3.1.0